Almost Done With Project!

+++ Source/WebContent/delete_assgn.jsp
@@ -13,11 +13,12 @@
 <body>
 	<%
 		Class.forName("com.mysql.jdbc.Driver");
-		Connection con = DriverManager.getConnection("jdbc:mysql://localhost:3306/SES?useSSL=false", "Manish", "Manny@123");
+		Connection con = DriverManager.getConnection("jdbc:mysql://localhost:3306/SES?useSSL=false", "Manish",
+				"Manny@123");
 		PreparedStatement pstatement = null;
 		Statement st = null;
-		
-		int i2[],i1=0,del;
+
+		int i2[], i1 = 0, del;
 		i2 = new int[10];
 		String s_name = request.getParameter("dltl");
 
@@ -25,25 +26,28 @@
 		try {
 			pstatement = con.prepareStatement(query);
 			pstatement.setString(1, s_name);
-		} catch (Exception e) {
 
-			session.setAttribute("error", e.getMessage());
-			response.sendRedirect("browse_word.jsp");
-		}
-		resultset = pstatement.executeQuery();
-		if (resultset.next()) {
+			resultset = pstatement.executeQuery();
+			resultset.next();
 			i1 = resultset.getInt("sid");
-		}
-		try
-		{
-		Statement statement = con.createStatement();
-		del= statement.executeUpdate("delete from assignment where sid =" + i1 + "");
-		session.setAttribute("error", "Assignment Deleted Successfully");
-		response.sendRedirect("browse_word.jsp");
-	
-		}
-		catch(Exception e)
-		{
+			Statement statement = con.createStatement();
+			
+			String queryCheck = "select count(*) as Total from assignment where sid = "+i1+"";
+			Statement stmt = con.createStatement();
+			ResultSet rs = stmt.executeQuery(queryCheck);
+			rs.next();
+			if(rs.getInt("Total") > 0){
+				
+				del = statement.executeUpdate("delete from assignment where sid =" + i1 + "");
+				session.setAttribute("error", "Assignment Deleted Successfully");
+						
+			}else{
+				session.setAttribute("error", "No Assignments To Delete");
+			}
+			
+			response.sendRedirect("browse_word.jsp");
+
+		} catch (Exception e) {
 			session.setAttribute("error", e.getMessage());
 			response.sendRedirect("browse_word.jsp");
 		}
